+++ iree/compiler/Codegen/Common/GPUDistributeSharedMemoryCopy.cpp
@@ -7,11 +7,11 @@
 #include <algorithm>
 #include <numeric>
 
-#include "iree/compiler/Codegen/Common/GPUUtils.h"
 #include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
 #include "iree/compiler/Codegen/PassDetail.h"
 #include "iree/compiler/Codegen/Passes.h"
 #include "iree/compiler/Codegen/Transforms/Transforms.h"
+#include "iree/compiler/Codegen/Utils/GPUUtils.h"
 #include "iree/compiler/Codegen/Utils/MarkerUtils.h"
 #include "mlir/Dialect/GPU/GPUDialect.h"
 #include "mlir/Dialect/Vector/Transforms/VectorTransforms.h"
